Move CountrySpecificPaymentField to own file
